+++ test/mocha/basic.test.js
@@ -1,28 +1,22 @@
 var chai = require('chai');
 var expect = chai.expect;
-var sinon  = require('sinon');
+// var sinon  = require('sinon');
 var fse = require('fs-extra');
-
+const {JsonConfig} = require('../../');
 const configObject = require('../config.json');
 const forkChild = require('../util/mocha_util').forkChild('json', configObject);
-var config = {configObject};
+const {schema} = require('./config');
 process.env.DELAY_TIME_BEFORE_EXIT = 1000;
-// var DELAY_TIME = 1000 + 100;
+
+var config = {configObject, schema};
 describe('basic test:',function() {
-    before(function() {
-        sinon.stub(process, 'exit');
-    });
-    after(function() {
-        sinon.restore();
-    });
+    it('succeeds if load necessary var ok', function (done) {
+        var settings = new JsonConfig(config);
 
-    it('succeeds if load necessary var ok', function () {
-        var settings = require('../..').getInstance(config);
-
-        var varstr = settings.loadNecessaryVar('var');
+        var varstr = settings.getValue('var');
         expect(varstr).to.be.exist;
-        var varFromMap = settings.getAllConfig().get('var');
-        expect(varFromMap).to.be.exist;
+        expect(settings.params).to.have.property('var');
+        done();
     });
 
     it('fail if can not load necessray var ',function(done) {
@@ -30,9 +24,9 @@ describe('basic test:',function() {
     });
 
     it('success if load necessary directory ok',function() {
-        var settings = require('../..').getInstance(config);
+        var settings = new JsonConfig(config);
         fse.ensureDirSync(configObject.dir);
-        var filepath = settings.loadNecessaryDirectory('dir');
+        var filepath = settings.getValue('dir');
         expect(filepath).to.be.exist;
     });
 
@@ -41,9 +35,9 @@ describe('basic test:',function() {
     });
 
     it('success if can load necessary integer',function() {
-        var settings = require('../..').getInstance(config);
+        var settings = new JsonConfig(config);
 
-        var value = settings.loadNecessaryInt('int');
+        var value = settings.getValue('int');
         expect(value).to.be.a('number');
     });
 
@@ -52,9 +46,9 @@ describe('basic test:',function() {
     });
 
     it('success if load necessary file',function() {
-        var settings = require('../..').getInstance(config);
+        var settings = new JsonConfig(config);
         fse.ensureFileSync(configObject.file);
-        var filepath = settings.loadNecessaryFile('file');
+        var filepath = settings.getValue('file');
         expect(filepath).to.be.exist;
     });
 
@@ -65,8 +59,8 @@ describe('basic test:',function() {
     });
 
     it('success if load necessary url',function() {
-        var settings = require('../..').getInstance(config);
-        var url = settings.loadNecessaryUrl('url');
+        var settings = new JsonConfig(config);
+        var url = settings.getValue('url');
         expect(url).to.be.exist;
     });
 
@@ -77,6 +71,105 @@ describe('basic test:',function() {
     it('fail if the loaded url not end with /',function(done) {
         forkChild('url', done);
     });
+
+    it('custom parse', function(done) {
+        const newConfig = {...config};
+        const SUFFIX = '_SUFFIX';
+        newConfig.schema = Object.assign({},schema, {
+            'object': {
+                type: {
+                    'key1': {type:String, required: true,}
+                },
+                custom: function({value}) {
+                    value.key1 = value.key1 + SUFFIX;
+                    return {error: null, value};
+                }
+            }
+        });
+        var settings = new JsonConfig(newConfig);
+        const obj = settings.getValue('object');
+        expect(obj.key1).to.be.equal(configObject.object.key1 + SUFFIX);
+        done();
+    });
+
+    it('sub custom parse', function(done) {
+        const newConfig = {...config};
+        const SUFFIX = '_SUFFIX_SUB';
+        newConfig.schema = Object.assign({},schema, {
+            'object': {
+                type: {
+                    'key1': {
+                        type:String, 
+                        required: true, 
+                        custom: function({value}) {
+                            value = value + SUFFIX;
+                            return {error: null, value};
+                        }
+                    }
+                },
+            }
+        });
+        var settings = new JsonConfig(newConfig);
+        const obj = settings.getValue('object');
+        expect(obj.key1).to.be.equal(configObject.object.key1 + SUFFIX);
+        done();
+    });
+
+    it('sub after parse', function(done) {
+        const newConfig = {...config};
+        newConfig.schema = Object.assign({},schema, {
+            'object': {
+                type: {
+                    'key1': {
+                        type:String, 
+                        required: true, 
+                    }
+                },
+                afterParse: function(value) {
+                    expect(value.key1).to.be.equal(configObject.object.key1);
+                    done();
+                }
+            }
+        });
+        new JsonConfig(newConfig);
+    });
+
+    it('preferred value', function(done) {
+        const newConfig = {...config};
+        const GIVEN_VALUE = 'ABC';
+        newConfig.schema = Object.assign({},schema, {
+            'keyx': {
+                type:String, 
+                required: true, 
+                preferred: GIVEN_VALUE,
+                afterParse: function(value) {
+                    expect(value).to.be.equal(GIVEN_VALUE);
+                    done();
+                }
+            }
+        });
+        new JsonConfig(newConfig);
+    });
+
+    it('preferred function value', function(done) {
+        const newConfig = {...config};
+        const GIVEN_VALUE = '123';
+        newConfig.schema = Object.assign({},schema, {
+            'keyy': {
+                type:String, 
+                required: true, 
+                preferred: function() {
+                    return GIVEN_VALUE;
+                },
+                afterParse: function(value) {
+                    expect(value).to.be.equal(GIVEN_VALUE);
+                    done();
+                }
+            }
+        });
+        new JsonConfig(newConfig);
+    });
+
 });