+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v3.2.0
+
+## Modify
+
+1. Remove the parameter of key from the callback of `AfterParseFunction`.
+
+# v3.1.0
+
+## Add
+
+1. Add preferred property to `Schema`, which used to give the default value of one element.
+
+# v3.0.0
+
+## Breaking changes
+
+1. Remove all the functions of `JsonConfig` and `ConsulConfigSync`.
+2. Delete the class of `ConsulConfig`.
+
+## Add
+
+1. Add the ability of wathing the changes from consul.
+2. Add the schema parameter to restrict the input from json object or consul server, it may been transformed to the type you desired.
+
 # v2.3.1
 ## Fix